enemy dies and refactor
