+++ mygame/src/kaboom.js
@@ -1,0 +1,11 @@
+import kaboom from "kaboom";
+
+export const k = kaboom({
+  global: true,
+  fullscreen: true,
+  scale: 1,
+  debug: true,
+  clearColor: [0, 0, 0, 1],
+});
+
+export default k;
